% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{est_ini}
\alias{est_ini}
\title{Initial estimation}
\usage{
est_ini(X, Y, quant = NA, method)
}
\arguments{
\item{X}{Array or DataFrame containing covariates}

\item{Y}{Array or DataFrame of outcomes}

\item{quant}{quantile for quantile estimation}

\item{method}{indicates the method to be used for M-estimation. Options include "mean", "quantile", "ols", "logistic", and "poisson".}
}
\value{
initial estimatior
}
\description{
\code{est_ini} function for initial estimation
}
