% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{optim_weights}
\alias{optim_weights}
\title{Gradient descent for obtaining the weight vector}
\usage{
optim_weights(
  j,
  X_lab,
  X_unlab,
  Y_lab,
  Yhat_lab,
  Yhat_unlab,
  w,
  theta,
  quant = NA,
  method
)
}
\arguments{
\item{j}{j-th coordinate of weights vector}

\item{X_lab}{Array or DataFrame containing observed covariates in labeled data.}

\item{X_unlab}{Array or DataFrame containing observed or predicted covariates in unlabeled data.}

\item{Y_lab}{Array or DataFrame of observed outcomes in labeled data.}

\item{Yhat_lab}{Array or DataFrame of predicted outcomes in labeled data.}

\item{Yhat_unlab}{Array or DataFrame of predicted outcomes in unlabeled data.}

\item{w}{weights vector POP-Inf linear regression (d-dimensional, where d equals the number of covariates).}

\item{theta}{parameter theta}

\item{quant}{quantile for quantile estimation}

\item{method}{indicates the method to be used for M-estimation. Options include "mean", "quantile", "ols", "logistic", and "poisson".}
}
\value{
weights
}
\description{
\code{optim_weights} function for gradient descent for obtaining estimator
}
