% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{psi}
\alias{psi}
\title{Esimating equation}
\usage{
psi(X, Y, theta, quant = NA, method)
}
\arguments{
\item{X}{Array or DataFrame containing covariates}

\item{Y}{Array or DataFrame of outcomes}

\item{theta}{parameter theta}

\item{quant}{quantile for quantile estimation}

\item{method}{indicates the method to be used for M-estimation. Options include "mean", "quantile", "ols", "logistic", and "poisson".}
}
\value{
esimating equation
}
\description{
\code{psi} function for esimating equation
}
