% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sim_data}
\alias{sim_data}
\title{Simulate the data for testing the functions}
\usage{
sim_data(r = 0.9, binary = FALSE)
}
\arguments{
\item{r}{imputation correlation}

\item{binary}{simulate binary outcome or not}
}
\value{
simulated data
}
\description{
\code{sim_data} function for the calculation of the matrix A
}
